sentimenter/pyproject.toml
[tool.poetry]
name = "sentimenter"
version = "0.1.0"
description = ""
authors = ["wayantrisna79 <wayantrisna79@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
